Prettify package manager name to help show Yarn 1 vs Berry
