+++ code/lib/create-storybook/src/commands/PreflightCheckCommand.ts
@@ -3,6 +3,7 @@ import {
   type JsPackageManager,
   JsPackageManagerFactory,
   PackageManagerName,
+  getPrettyPackageManagerName,
   invalidateProjectRootCache,
 } from 'storybook/internal/common';
 import { CLI_COLORS, deprecate, logger } from 'storybook/internal/node-logger';
@@ -17,15 +18,6 @@ export interface PreflightCheckResult {
   packageManager: JsPackageManager;
   isEmptyProject: boolean;
 }
-
-/** Human-friendly labels for each package manager type */
-const PACKAGE_MANAGER_LABEL: Record<PackageManagerName, string> = {
-  [PackageManagerName.NPM]: 'npm',
-  [PackageManagerName.YARN1]: 'Yarn Classic (v1)',
-  [PackageManagerName.YARN2]: 'Yarn Berry',
-  [PackageManagerName.PNPM]: 'pnpm',
-  [PackageManagerName.BUN]: 'Bun',
-};
 
 /**
  * Command for running preflight checks before Storybook initialization
@@ -72,9 +64,7 @@ export class PreflightCheckCommand {
       force: options.packageManager,
     });
 
-    logger.info(
-      `Package manager: ${PACKAGE_MANAGER_LABEL[packageManager.type] ?? packageManager.type}`
-    );
+    logger.info(`Package manager: ${getPrettyPackageManagerName(packageManager.type)}`);
 
     // Install base project dependencies if we scaffolded a new project
     if (isEmptyDirProject && !options.skipInstall) {
